Some settings are not meant to be overridden.  Refuse to allow overriding the DBSchemaVer setting.
If we decide to do the same for other settings, we can easily move the approval to a separate function, but since (for now) we only have one disallowed override, this just adds the check inline.  The check is performed in MythDB::OverrideSettingForSession() rather than in each application's main(), so the user will see log output such as:

2010-02-05 15:39:21.211 Setting 'DBSchemaVer' being forced to '1252'
2010-02-05 15:39:21.212 ERROR: Refusing to allow override for 'DBSchemaVer'.

(where the initial "Setting...being forced" comes from main()), but this should still be sufficient to alert the user to the fact that the override they specified is not allowed and was ignored.


git-svn-id: http://svn.mythtv.org/svn/trunk@23483 7dbf422c-18fa-0310-86e9-fd20926502f2