Start of really porting MythCenter-wide to MythUI.
There are still a number of screens needing attention, and the color selection for selected text needs to be improved. What this does is reduce the number of fonts used from four to two, name fonts according to the state they represent (instead of the color used for that state in default-wide/base.xml), along with layout tweaks for a few frequently used screens.

The fonts used are 'Liberation Sans' for large text and 'Droid Sans' for small text. Droid Sans is closer in resemblance to the original font used in this theme and is very readable at small font sizes, but it looks bad at large font sizes. I believe the problem at large font sizes is with the kerning, but I haven't really investigated it. Both are designed by the same guy and both are freely redistributable fonts available in easy to install distro packages.

git-svn-id: http://svn.mythtv.org/svn/trunk@23487 7dbf422c-18fa-0310-86e9-fd20926502f2