FFmpeg sync to revision 21657, Refs #7233
The most important change in this sync is probably the optimized
H.264 decoder (approx. 15% faster on X86_64 linux). For fans of hardware
assisted decoding VDPAU mpeg4 asp and DXVA2 (H.264, VC1, WMV3) were added
(MythTV needs to be changed to make use of them). Support for BluRay
subtitles and PCM are the most prominent added decoders.



git-svn-id: http://svn.mythtv.org/svn/trunk@23525 7dbf422c-18fa-0310-86e9-fd20926502f2