Fixes MythUI XML parsing error reports.
Two changes. First, info is passed to the parser so that it can report not just the line of the error, but also what file it was in. Second, a showWarnings bool is added for reporting duplicate definitions only when loading the first theme xml file in the search path that successfully loads.

git-svn-id: http://svn.mythtv.org/svn/trunk@23539 7dbf422c-18fa-0310-86e9-fd20926502f2