Prettify MythCenter-wide guidegrid at 1920x1080.
This adds a column so that five rather than four half hour slots are shown on screen at once. The prettification is specific to 1920x1080 because buttonlist's and the grid both do not scale correctly to different resolutions so alignment between the buttons and grid gets lost when this screen is scaled.

git-svn-id: http://svn.mythtv.org/svn/trunk@23549 7dbf422c-18fa-0310-86e9-fd20926502f2