Fix a problem with the new MythSignalingTimer consuming all memory and causing both MythFrontend and MythWelcome to become unresponsive and eventually to be killed by the OOM killer.
The problem is triggered when the main UI thread is stopped waiting for another
external process to finish, e.g. when using an external player or when starting
the FE from MythWelcome. Because the timer is run in its own thread it
continues to emit the timeout signals which just get queued up waiting to be
handled.

This patch just start/stops the timer when SetDrawEnabled(true/false) is
called and changes myth_system() to disable drawing when an external process is
running and re-enables it when it finishes.


git-svn-id: http://svn.mythtv.org/svn/trunk@23575 7dbf422c-18fa-0310-86e9-fd20926502f2