If a channel icon cannot be loaded in the EPG don't keep trying to find it. We still try to find each icon once so the EPG is still a little sluggish but is usable.
I'm in two minds as to whether we should nuke the icon setting in the DB if
an icon cannot be found but that could be annoying if there is a temporary
network glitch and all your carefully selected icon settings get wiped.

I think the real solution is for the users seeing this to fix their broken
setups by making sure any icon referenced in the DB exists and is readable
by the user running MythFrontend or MythBackend for remote FE's.

Fixes #7929.

git-svn-id: http://svn.mythtv.org/svn/trunk@23579 7dbf422c-18fa-0310-86e9-fd20926502f2