Since Qt4.2, Qt has supported loading fonts that are not installed on the system, so now we can load theme-provided fonts for UI themes.
This change adds a new class, MythFontManager, to manage font file loading and adds calls to ensure we scan theme directories for font files before loading the theme and that we unload no-longer-necessary font files when unloading the theme.

For themers, simply ensure you place (redistributable) TrueType font (.ttf), OpenType font (.otf), or TrueType font collection (.ttc) files anywhere within the theme directory structure, specify the font face within the theme XML as before, and the fonts will be used--whether they're installed on the system or not.  No changes to theme files are required.

Thanks to Chris Pinkham for tracking down the Qt QFontDatabase class and its ability to load application fonts--for providing the idea on which this patch is based.


git-svn-id: http://svn.mythtv.org/svn/trunk@23594 7dbf422c-18fa-0310-86e9-fd20926502f2