Fixes #8006. Refs #7195. Fixes some regressions in preview generation. Most importantly this addresses the segfault. But it also removes some of the cruft in the PreviewGenerator constructor that had built up to deal with places that called the preview generator without setting pathname as per the PreviewGenerator documentation. This means we aren't doing the very expensive StorageGroup search quite as many times.
This also saves the pathname after doing a StorageGroup search in the PlaybackBoxHelper.

There are still a number of ways left to speed up the PreviewGenerator: For instance, using fork instead of running an external program to get fault tolerance, which will also address #7136. We also do not need to open a new command socket if we convert the backend preview generation to use a non-blocking call and just send a MythEvent when we're done. But I wanted to limit this commit to really only fix regressions since we're in feature freeze (the pathname saving is pbbh is a bit of a stretch, but we actually did do this before in a different way).

git-svn-id: http://svn.mythtv.org/svn/trunk@23599 7dbf422c-18fa-0310-86e9-fd20926502f2