Carry over [23602] from trunk.
Revert part of [18792].  That commit broke the delete code in that the
delete lock was no longer released before the truncating started.
This means that if a second recording was 'deleted' before the first
had finished truncating, then the second recording's unlink would
block until after the first truncate was finished.  If the backend
died before the first truncate finished, the second file would not
be deleted since it had never been unlinked.  This commit restores
the original functionality so that all files are allowed to be
unlinked and then the truncates happen sequentially in the background.

This commit also provides the same functionality for non-recording
files that are being deleted using truncating delete.  The
truncates are no longer being protected by the deletelock, now the
truncate is only protected by the truncate_and_close_lock, so a
long-running truncate does not hold up other files from being
unlinked.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@23604 7dbf422c-18fa-0310-86e9-fd20926502f2