Allow the 'wants' events' field in the ANNounce string to specify whether a client wants to receive (0) no events, (1) all events, (2) non-system events, or (3) only system events.
This is the basis for a fix for issues encountered when
MythWeb doesn't like receiving the "client connected"
system event when MythWeb itself connects to the master
backend.

Also modifies the Perl bindings to allow them to specify
this value when connecting since I used the Perl bindings
to help test this functionality.

References #8084, and #8117.



git-svn-id: http://svn.mythtv.org/svn/trunk@23636 7dbf422c-18fa-0310-86e9-fd20926502f2