Clean up inuseprogram handling a bit.
This makes two substansive changes, 1st when we delete an inuseprogram entry we reset the lastInUseTime, 2nd we only update lastInUseTime when a inuseprogram insert or update works. If for some reason an update fails these two changes make sure that we don't squash subsequent attempts to update the entry. This also adds a DEBUG_IN_USE define which prints additional debugging, disabled by default, and it sets the usedFor strings in all MarkAsInUse calls from the TVRec class.

I've seen cases where the inuseprograms isn't updated until we're 15 minutes into the program, i.e. the first time an update will take. I haven't been able to reproduce this reliably enough to be sure this fixes the bug, but at the very least it should decrease the severity of the problem.

git-svn-id: http://svn.mythtv.org/svn/trunk@23668 7dbf422c-18fa-0310-86e9-fd20926502f2