Allow a user style sheet to be loaded in MythUIWebBrowser.
Also add a default css file that just changes the color of the background of
the focused link to make it more visible when trying to select a link using
the keyboard.

Themers can if they wish use their own default css file which will be used
anywhere the web browser widget is used or add a <userstylesheet> tag to
override the default and use a different one.

Closes #8129.


git-svn-id: http://svn.mythtv.org/svn/trunk@23714 7dbf422c-18fa-0310-86e9-fd20926502f2