Fix MythXML preview generation for non-default time offsets.
Fixes MythXML preview generation when a non-default time offset is specified, which resulted in filename-handling issues after the recent changes to ProgramInfo path handling.  Patch was inspired by a patch from ThisOtherGuy in IRC, with modifications by me to handle additional (pre-existing) issues with the filename handling I noticed while reviewing his patch.  Thanks to ThisOtherGuy for identifying the bug and for providing the initial patch.

This is the MythXML bug that caused the issue reported by #8118.  That issue has since been fixed by modifying MythWeb to use a more-appropriate approach to request previews.  Refs #8118.


git-svn-id: http://svn.mythtv.org/svn/trunk@23764 7dbf422c-18fa-0310-86e9-fd20926502f2