Fixes fullscreen window mode.
The check for whether to make the window fullscreen added in [23578] was not working properly since the "real" (full-screen-sized) values for screen width and height were set by the time GetScreenSettings() was called, so they were no longer the full screen flag value, 0.  This changes the check to rely, instead, on the settings values, and also adds some logging output to make debugging future issues easier.

Refs #7996.  Refs #8021 (though the change may not fix Compiz, mythfrontend should now properly request fullscreen mode when appropriate).  Thanks to stuarta and EnderTheThird for reporting the issue.


git-svn-id: http://svn.mythtv.org/svn/trunk@23765 7dbf422c-18fa-0310-86e9-fd20926502f2