Populate the alsa devices in audio setup by using the API, rather than a static list of devices that are very likely not to apply on many/most systems.  Also has the advantage of providing a "plain-english" description of the ALSA node selected in the tooltip. More or less, the ALSA device contents in the pulldown will now match the output of aplay -L.
Also supports definitions in the ~/.asoundrc file, so long as the hint{} attribute was used.  For an example, see http://www.mythtv.org/wiki/Configuring_Digital_Sound .

Very sincere thanks to Daniel K, Stuart M, and Mike D. who gave me a lot of assistance in getting this right and cleaning it up.


git-svn-id: http://svn.mythtv.org/svn/trunk@23794 7dbf422c-18fa-0310-86e9-fd20926502f2