Fixes #6662.  Fixes editable comboboxes.
Since [17978], every time the user changed the text in the combobox (for each keystroke entered), a new entry was added to the combobox list of values.  This change ensures that there's only one entry in the box corresponding to the user-edited content.  Thanks to mythtv at dadeos for the patch.

Unfortunately, this changes a low-level header, so a lot of recompilation will be required, though a reconfigure shouldn't be necessary.


git-svn-id: http://svn.mythtv.org/svn/trunk@23803 7dbf422c-18fa-0310-86e9-fd20926502f2