Fix an off by 1 error in MythUIText::cutDown
When asking QFontMetrics::width for the width needed for the given
text, the last character of the text was not being included. If that last
character happened to be the difference between the text fitting or not, we
would incorrectly conclude that it did fit.

Note: this should have gone in as part of r23790


git-svn-id: http://svn.mythtv.org/svn/trunk@23811 7dbf422c-18fa-0310-86e9-fd20926502f2