Fixes MythWeather setup segfault.
The fix for #8169 assumed that the MythWeather plugin was running when it changed the SourceManager::findScripts() busy popup to appear on the weather stack.  When entering MythWeather setup through Utilities/Setup|Setup|Info Center Settings|Weather Settings|Screen Settings, this resulted in a NULL MythScreenStack.  Therefore, if the MythScreenStack for the weather stack is null, just use the popup stack.

Fixes #8223.  Refs #8169.


git-svn-id: http://svn.mythtv.org/svn/trunk@23817 7dbf422c-18fa-0310-86e9-fd20926502f2