Fixes #7832.  Set the supplemental group IDs appropriately when changing the process persona for the --user command-line argument to mythbackend.
This change uses a slightly simplified version of Mario's patch on the ticket and puts the code that sets the supplemental group IDs after the call to setgid() to be slightly more portable (even though none of the platforms on which MythTV runs would have had issues with the other approach).

Thanks to Mario L. for the work on the --user command-line argument.


git-svn-id: http://svn.mythtv.org/svn/trunk@23835 7dbf422c-18fa-0310-86e9-fd20926502f2