Backport [23790] and [23811] from trunk.
Fix an off by 1 error in MythUIText::cutDown.

When calculating the area needed for the given text, the last character of
the text was not being included. If that last character happened to be the
difference between the text fitting or not, we would incorrectly conclude
that it did fit.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@23837 7dbf422c-18fa-0310-86e9-fd20926502f2