EITFixup: heuristic fixup for short events marked as category type "movie"
RTL and finish channels have events for tvshows/series marked as movies.
There is unfortunately no other indication in the description so we use
the time as heuristic. Only event longer than 75 minutes are considered
to be movies. All other events are classified as tvshows.
This makes the canned movie search useable on those channels.
Fixes #8216, Refs #6844

Based on a patch by: dekarl spaetfruehstuecken org

git-svn-id: http://svn.mythtv.org/svn/trunk@23973 7dbf422c-18fa-0310-86e9-fd20926502f2