Moves the command line parsing for mythbackend into MythCommandLineParser and moves the helper functions into main_helpers.cpp.
Note: This is in support of moving the backend config from mythtv-setup into the backend webserver. Many of the functions performed once now will be within a loop so backend functionality can be reinitialized after it is reconfigured without needing to manually restart any of the backends. Hence the desire to simplify main().

git-svn-id: http://svn.mythtv.org/svn/trunk@24030 7dbf422c-18fa-0310-86e9-fd20926502f2