NVR: handle modified frame dimensons from V4L2 devices
The V4L2 set format ioctl is allowed to change the requested
frame dimensions if it doesn't support a capturing resolution.
This moves the V4L2 set format infront of buffer and codec
initialization so that we don't have to handle frame size changes.
Fixes analog recordings with the Auvitek AU8522 found in the
Hauppauge WinTV-HVR-950Q which doesn't support our default
resolution of 480x480.
Tested with a modified bttv kernel module.

git-svn-id: http://svn.mythtv.org/svn/trunk@24063 7dbf422c-18fa-0310-86e9-fd20926502f2