Fix a deadlock in TV::SwitchSource() that occured when using the NEXTSOURCE/PREVSOURCE key bindings to switch sources.
The code in TV::timerEvent() was holding a lock, causing a deadlock when TV::SwitchSource() requested the PlayerContext.  This changes TV::SwitchSource() to work like the related TV::SwitchCards() and TV::ToggleInputs functions by using the PlayerContext that TV::timerEvent() passes through TV::ProcessKeypress() and, finally, to TV::ActivePostQHandleAction(), which calls TV::SwitchSource().

This seems to have been broken in the mythtv-vid branch, but the code was unreachable/untestable due to broken key bindings (which were fixed for #6472).

Refs #6472.  Fixes #8190.  Thanks to jafa for finally getting a backtrace.


git-svn-id: http://svn.mythtv.org/svn/trunk@24274 7dbf422c-18fa-0310-86e9-fd20926502f2