Carry over [24278] from trunk.
Add a new LIVETV_STARTED System Event to signal when LiveTV starts.  We can't
easily send a PLAY_STARTED with the proper program info, and LIVETV_STARTED
helps differentiate between playback of a recording vs LiveTV.

Send a PLAY_CHANGED System Event when we jump to the next program in
NuppelVideoPlayer.

References #8388.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@24279 7dbf422c-18fa-0310-86e9-fd20926502f2