Internal connection classes are now DBConnection, BEConnection, and XMLConnection, with respective caches being BECache and DBCache. The top level classes will remain as they currently are, with MythDB, MythBE, and MythXML. Additionally, the 'type' input for backend connections is replaced with a 'noshutdown' boolean, to indicate whether the backend is allowed to shutdown while the connection is active.
The backend connection is now non-blocking, and supports events.  This is handled through a new BEEvent class. Two new boolean inputs are available: 'generalevents' and 'systemevents'. Event handling is done by registering a regular expression to match against, and a method to be called when a match is found.

Events can be registered to a BEConnection object manually, or by subclassing the BEEvent class. Two examples of this are the BEEventMonitor class, which registers a simple 'catch all' event and prints all messages to screen for debugging purposes, and the FileTransfer class, which will now listen for UPDATE_FILE_SIZE events if used to open a recording.


git-svn-id: http://svn.mythtv.org/svn/trunk@24292 7dbf422c-18fa-0310-86e9-fd20926502f2