Add a helper method so that we can check whether we are in the UI thread or not.  We don't want to block LIRC, the joystick, or drawing when called outside the UI thread so use this helper to check to see if we are in the UIthread.  Disabling drawing from the UI thread can also cause segfaults as indicated in #8194 and #8391.
References #8194 and #8391.

This fix is only temporary in trunk.  I have a different version in a
local checkout that will be committed later as part of a patch to move
parts of MythContext to libmythdb where they can be utilized by
myth_system() and the rest of libmythui.

This does update the API version, so make clean, etc..



git-svn-id: http://svn.mythtv.org/svn/trunk@24310 7dbf422c-18fa-0310-86e9-fd20926502f2