Carry over [24310] from trunk.
Add a helper method so that we can check whether we are in the UI thread or
not.  We don't want to block LIRC, the joystick, or drawing when called
outside the UI thread so use this helper to check to see if we are in the
UIthread.

Fixes #8194 and #8391.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@24311 7dbf422c-18fa-0310-86e9-fd20926502f2