Backport [24179].  It's purely cosmetic, but it makes scheduler dumps much easier to read so I'd like to have it in the release.
Try to make the verbose scheduler output a little easier to read.
First, print a legend for the cryptic punctuation characters prefixed
to each line as the scheduling algorithm is traced.  Second, shorten
the title/subtitle column as needed when a prefix is used to keep
everything else aligned instead of shifting the entire line over.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-23-fixes@24431 7dbf422c-18fa-0310-86e9-fd20926502f2