Compilation fix for MS Windows.  Do not allow the --user option on Windows.  Refs #8121.
This is an updated version of the patch submitted by Jeff Lu, taking into account changes in [24030].  I can not test the fix on Windows, and since it's not the same patch submitted by Jeff, any breakage is due to me.  So if it doesn't work, please let me know with a message to the mailing list.


git-svn-id: http://svn.mythtv.org/svn/trunk@24482 7dbf422c-18fa-0310-86e9-fd20926502f2