lavf: fix evob (HD DVD) probing with small probe buffers
The mpeg ps probe function needs for some evob 1M probe buffer since it
detects many invalid packets due to PES start code emulation in audio
and private streams. If the PES payload is skipped for this streams the
emulated start codes are not found and the much smaller probe buffer can
be used. Refs #7233

git-svn-id: http://svn.mythtv.org/svn/trunk@24617 7dbf422c-18fa-0310-86e9-fd20926502f2