Short answer:
If your patch breaks, ESC : %s/gContext/gCoreContext/g ENTER in 99% of cases.
If that doesn't fix it, it might be s/gContext->GetMainWindow/GetMythMainWindow/g
And make clean since the API version changed.

Long answer:

This commit is a MythContext reorganization patch.  The main goal is to get
the non-GUI portions of MythContext deeper into the library dependency
tree.  libmyth depends on libmythui which depends on libmythdb.  This means
that neither libmythdb or libmythui can use anything in the current
MythContext.  This patch addresses this issue by moving most of MythContext
into libmythdb as MythCoreContext.  This will for instance allow libmythdb
and libmythui code to use the existing master backend connection rather than
having to open a new connection to talk to the master.

To accomplish the above, this patch...

- moves the non-GUI portions of MythContext to a new MythCoreContext class
  located in libs/libmythdb/mythcorecontext.(cpp|h).  Also changes
  #include mythcontext.h to #include mythcorecontext.h where applicable.

- adds a new gCoreContext global pointer to the MythCoreContext instance.
  gCoreContext is now used in most places that gContext was used previously
  since most of these methods are now in MythCoreContext.  The global
  gCoreContext creation/deletion is handled by the global MythContext.
  MythContext does not inherit from MythCoreContext because we need the
  routines in MythCoreContext to be accessible via the gCoreContext pointer
  inside libmythdb and libmythui.

- moves mythsystem.(cpp|h) from libmythui to libmythdb

- moves the following methods out of MythContext:

  MythContext::sendPlaybackStart() -> sendPlaybackStart() in libmyth/util.cpp
  MythContext::sendPlaybackEnd()   -> sendPlaybackEnd() in libmyth/util.cpp

- gets rid of the following methods:

  MythContext::GetMainWindow()     (use mythmainwindow.h's GetMythMainWindow())
  MythContext::SetMainWindow()     (not needed without ::GetMainWindow())
  MythContext::TranslateKeyPress() (already commented as deprecated,
                                    use GetMythMainWindow()->TranslateKeyPress())

- creates the following method(s):

  MythContext::SetDisableEventPopup(bool) (used by frontend's exit prompt
                                           code to disable popups since
                                           MythContext::SetMainWindow() does
                                           not exist anymore)

- performs some minor header cleanup, converting some files from #including
  mythcontext.h to instead include mythverbose.h since all they needed was
  access to the VERBOSE macro.


You'll need to do a make clean on this one.  The binary API is updated and
libmythdb, libmythui, and libmyth are all changed more than a tiny bit.
Most of the rest of the changes are s/gContext/gCoreContext/g with a little
s/mythcontext.h/mythcorecontext.h/ thrown in for #include changes.

=============================================================================



git-svn-id: http://svn.mythtv.org/svn/trunk@24623 7dbf422c-18fa-0310-86e9-fd20926502f2