Remove the xdg-screensaver support by reverting [24559].
While the promise of xdg-screensaver is great, it seems that neither distros nor desktop environments have any desire to support it.  Users on various distros with various desktop environments using various (including supported and not supported) screensaver applications found that it did not work for them.  Until it's properly supported by distros and desktop environments, it's not useful.

Fixes #8443.  Refs #8357.  Refs #7291.


git-svn-id: http://svn.mythtv.org/svn/trunk@24636 7dbf422c-18fa-0310-86e9-fd20926502f2