Ensure we always clean up properly when exiting mythcommflag.
This adds a cleanup guard to mythcommflag to ensure gContext is always deleted, regardless of where we exit (as we exit in different places for many of the command-line arguments).  There were several locations where we failed to delete the MythContext on exit, and--coupled with recent changes to the event-passign code--caused problems in some of these situations, dependent on what happened as we were exiting.  Refs #8310.


git-svn-id: http://svn.mythtv.org/svn/trunk@24646 7dbf422c-18fa-0310-86e9-fd20926502f2