Ensure unique programid generation for XMLTV-based grabbers that provide episode and season numbers.
The historical encoding of episode and season numbers into programids (since [5248])--appending unpadded episode and season numbers to the end of the seriesid--could cause non-unique programid's (for example, episode 11 of season 2 and episode 1 of season 12 were both represented with the same "112").  Since changing the programid would cause duplicate matching to fail against previously-recorded episodes, we have to maintain the approach of appending unpadded episode and season numbers to the seriesid.

To ensure uniqueness of generated programids, this change requires that season always be represented as a single character.  To increase the number of seasons for which we can generate a programid, the season is represented in base 36.  If the season can not be represented as a single character, a programid is not generated, so the scheduler will fall back to using normal duplicate-matching methods (title/subtitle/description).

This change will cause duplicate matching to fail for any episodes in seasons 10 or above for users whose XMLTV grabbers provide episode and season numbers--meaning that episodes of high-numbered seasons will be re-recorded.  All episodes in low-numbered seasons will be unaffected by the change.  However, after this change, users will no longer have the problem that episode 1 of seasons 11 through 19 are not recorded because the user previously recorded episode 11 of seasons 1 through 9, etc.  A false negative on duplicate matching is better than a false positive, and we have to change some programids to fix the bug.

This change does not attempt to make the episode and season information in the programid parseable values.  Parsing of episode and season information from programid is discouraged as there are probably more users for which the episode and season information does not appear in the programid than users for which it does.



git-svn-id: http://svn.mythtv.org/svn/trunk@24724 7dbf422c-18fa-0310-86e9-fd20926502f2