MythNetvision/Core RSS Parsing: Add support for a <mythtv:customhtml> tag.  Enables a grabber to indicate that it is capable of producing custom HTML to produce embedded video, mashups, etc.  Parsed but not yet used, but a requirement for some grabbers when proxied through the backend.  Which brings me to my next point...
Add a GetInternetSearch method to MythXML.  Allow for the backend to respond to grabber search requests.  This is a move towards having all installed internet content scripts being managed by the backend.  Once can now do:

http://BackendIP:6544/Myth/GetInternetSearch?Grabber=bbciplayer.py&Query=doctor%20who&Page=1 

To get the XML page 1 of search results for Doctor Who with the bbciplayer.py grabber.  Not yet used by MNV, but it will be in the next few weeks.

Bumps the library version again.  Plugins->recompile();


git-svn-id: http://svn.mythtv.org/svn/trunk@24777 7dbf422c-18fa-0310-86e9-fd20926502f2