Fix the DB schema version check.
The most important part of this change is to prevent an app running in a non-interactive session (mythbackend) from running when the DB schema version is newer than the version that the application understands.  This check was broken sometime before 0.23 was released, and mythbackend's running against a newer-than-understood version has been causing confusion for users of a certain distro whose package scripts don't properly restart MythTV applications after an upgrade.

This also modifies the order of operations such that we don't back up the database if the schema version is newer than the understood version--i.e. if we're just going to shut down and not use the database.


git-svn-id: http://svn.mythtv.org/svn/trunk@24820 7dbf422c-18fa-0310-86e9-fd20926502f2