Merge libmythui changes from the libmythui-osd branch.
In summary:

- Add 2 new hardware rendering contexts, MythRenderOpenGL and
MythRenderD3D9 (MythRenderVDPAU was added prior to the 0.23 release).
- Add 3 new theme painters, MythD3D9Painter, MythQImagePainter (for
rendering directly to a QImage) and MythYUVAPainter (a sub-class of
MythQImagePainter that converts all rendering to the YUV colorspace).
- Add 1 new MythUI widget, MythUIEditBar, which will be used to add
themed, edit functionality to the OSD (there is no reason why it cannot
be used in the main UI).
- Make various improvements to the painter classes and convert them to
use the new hardware rendering contexts where appropriate.
- Sundry small changes to other libmythui classes to better facilitate
the new OSD rendering code.

There should be no visual difference in the rendering of the main UI
using either the Qt or OpenGL painters.

Direct3D9 rendering can be enabled using 'mythfrontend -O
ThemePainter=d3d9' but I have not added this to the settings as I am
still unsure whether it is worthwhile having both Direct3D9 and OpenGL
rendering.

Refs #7650


git-svn-id: http://svn.mythtv.org/svn/trunk@24822 7dbf422c-18fa-0310-86e9-fd20926502f2