Fix key binding initialization after initial database creation.
When the database does not exist, key bindings are not written to the database.  Therefore, the Global, Browser, and Main Menu key bindings were not properly initialized when mythtv-setup or mythfrontend was used to create the initial database.  Upon restarting either program after creating the initial database, the bindings were properly initialized.

This change moves key initialization in MythMainWindow to a separate function and adds a function to reset key bindings, and calls that function from mythfrontend or mythtv-setup after the database upgrade check.  These changes will also be used with an upcoming change to allow users to reset keybindings to default values in MythControls/Edit Keys.

The change also allows simplification of the 1256 DB update, since the key-reloading hack is no longer required after this fix.

There is a binary version change, so you'll likely need a make clean and plugins will need to be recompiled.  (Word has it that another binary version change is coming very shortly, so you might want to wait a few minutes before updating and recompiling.)



git-svn-id: http://svn.mythtv.org/svn/trunk@24839 7dbf422c-18fa-0310-86e9-fd20926502f2