Remove the "Independent Muting of Left and Right Audio Channels" setting and add a key binding CYCLEAUDIOCHAN that provides the same functionality.  This approach allows users in areas where independent muting of audio channels to use either or both MUTE and CYCLEAUDIOCHAN (allowing them to easily mute all channels without having to cycle through them all to get to full mute).
Note that the actual functionality of "Independent Muting of Left and Right Audio Channels" has been broken for quite some time (or at least doesn't work with my configuration).  This was true before the recently-committed audio code refactor and before this patch.  This patch doesn't change that--it simply changes where/how you request the functionality.



git-svn-id: http://svn.mythtv.org/svn/trunk@24869 7dbf422c-18fa-0310-86e9-fd20926502f2