Improved display of DVD subtitles in containers other than the original.
This changes the default ffmpeg colour choice of yellow to white, so
that we get greyscale rendering rather than shades of yellow and fixes
the range so that we get the full scale from black to white.

This is by no means a complete fix as the palette is still a best guess
and hence the edges are sometimes brighter than the text and sometimes
the brightest shade (i.e. white) is not used - and hence the subtitles
appear grey.

Closes #8222


git-svn-id: http://svn.mythtv.org/svn/trunk@24956 7dbf422c-18fa-0310-86e9-fd20926502f2