Add Generic Metadata Classes.  This adds a "universal" method of dealing with video, music, and game metadata to libmythui (where it lives pending my writing an all-purpose online imagery browser).
In short, these classes allow the plugin or frontend developer to easily add metadata lookups for any kind of data to any type of screen.  The workflow is as follows:

1) Create a pointer to a new MetadataLookup object.
2) Use the various "sets" from metadatacommon.h to populate as much information as is necessary to complete a metadata grab. (Title, Subtitle, Season, Episode, TMDB #)
OPTIONAL 2.5) lookup->SetAutomatic(true) to tell the query thread to attempt fuzzy name matching.  By default this matching is quite strict to avoid false positives.
3) Pass the semi-populated lookup object to the MetadataDownload thread.
4) Listen for MetadataLookupEvents, which can pass back a single or list of matching lookups.  If you get multiple lookups, pick the one you want and chuck it back into the queue.
5) When the final lookup arrives, it is a fully populated MetadataLookup object.  Use the "Gets" to pull out all data relevant to your media type.
OPTIONALLY:
6) The lookup returns a map of all available artwork for the item.  Construct another metadatalookup, create a DownloadMap of the artwork you want downloaded, and add it to the MetadataImageDownload thread.
7) Wait for an ImageDownloadEvent, which will return the same downlload map pointing at the local locations of all your requested files.  Supports local and SG image download.

There are a LOT of cool things that this can do, and the above only scratches the surface.  I am sure there are also parts of this that need work and could be cleaner, we'll try to get those ironed out well in advance of .24.

This moves the functional Movie and Television grabbers to trunk/mythtv/programs/scripts/metadata in anticipation of their use in recordings in the very near future.  It also svn del's the secondary grabbers which are no longer functional using this scheme.  If you would like to see support for KinoX, Allocine, or Alpacine returned, please feel free to contact me for help in writing a grabber to conform to the new spec.  Since there is only one choice for each of these grabbers yet, I have not moved the metadata script picker into the core yet.


git-svn-id: http://svn.mythtv.org/svn/trunk@24960 7dbf422c-18fa-0310-86e9-fd20926502f2